+++ vite.config.js
@@ -3,6 +3,12 @@ import { defineConfig } from 'vite';
 export default defineConfig({
   root: 'src',
   build: {
+    rollupOptions: {
+      input: {
+        index: 'src/index.html',
+        reports: 'src/reports.html'
+      }
+    },
     outDir: '../dist',
     emptyOutDir: true
   },